# swagger.yaml
swagger: "2.0"
info:
  description: Conduit API
  version: 1.0.0
  title: Conduit API
  contact:
    name: RealWorld
    url: "https://realworld.io"
  license:
    name: MIT License
    url: "https://opensource.org/licenses/MIT"
basePath: /api
schemes:
  - https
  - http
produces:
  - application/json
consumes:
  - application/json
securityDefinitions:
  Token:
    description: >
      For accessing the protected API resources, you must have received a a
      valid JWT token after registering or logging in. This JWT token must then
      be used for all protected resources by passing it in via the
      'Authorization' header.


      A JWT token is generated by the API by either registering via /users or
      logging in via /users/login.


      The following format must be in the 'Authorization' header :

          Token: xxxxxx.yyyyyyy.zzzzzz

    type: apiKey
    name: Authorization
    in: header
paths:
  /users/login:
    post:
      summary: Existing user login
      description: Login for existing user
      tags:
        - User and Authentication
      operationId: Login
      parameters:
        - name: body
          in: body
          required: true
          description: Credentials to use
          schema:
            $ref: "#/definitions/LoginUserRequest"
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/UserResponse"
        "401":
          description: Unauthorized
        "422":
          description: Unexpected error
          schema:
            $ref: "#/definitions/GenericErrorModel"
  /users:
    post:
      summary: Register a new user
      description: Register a new user
      tags:
        - User and Authentication
      operationId: CreateUser
      parameters:
        - name: body
          in: body
          required: true
          description: Details of the new user to register
          schema:
            $ref: "#/definitions/NewUserRequest"
      responses:
        "201":
          description: OK
          schema:
            $ref: "#/definitions/UserResponse"
        "422":
          description: Unexpected error
          schema:
            $ref: "#/definitions/GenericErrorModel"
    get:
      summary: Get current user
      description: Gets the currently logged-in user
      tags:
        - User and Authentication
      security:
        - Token: []
      operationId: GetCurrentUser
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/UserResponse"
        "401":
          description: Unauthorized
        "422":
          description: Unexpected error
          schema:
            $ref: "#/definitions/GenericErrorModel"
    put:
      summary: Update current user
      description: Updated user information for current user
      tags:
        - User and Authentication
      security:
        - Token: []
      operationId: UpdateCurrentUser
      parameters:
        - name: body
          in: body
          required: true
          description: User details to update. At least **one** field is required.
          schema:
            $ref: "#/definitions/UpdateUserRequest"
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/UserResponse"
        "401":
          description: Unauthorized
        "422":
          description: Unexpected error
          schema:
            $ref: "#/definitions/GenericErrorModel"
...
definitions:
  LoginUser:
    type: object
    properties:
      email:
        type: string
      password:
        type: string
        format: password
    required:
      - email
      - password
  LoginUserRequest:
    type: object
    properties:
      user:
        $ref: "#/definitions/LoginUser"
    required:
      - user
  NewUser:
    type: object
    properties:
      username:
        type: string
      email:
        type: string
      password:
        type: string
        format: password
    required:
      - username
      - email
      - password
  NewUserRequest:
    type: object
    properties:
      user:
        $ref: "#/definitions/NewUser"
    required:
      - user
  User:
    type: object
    properties:
      email:
        type: string
      token:
        type: string
      username:
        type: string
      bio:
        type: string
      image:
        type: string
    required:
      - email
      - token
      - username
      - bio
      - image
  UserResponse:
    type: object
    properties:
      user:
        $ref: "#/definitions/User"
    required:
      - user
  UpdateUser:
    type: object
    properties:
      email:
        type: string
      token:
        type: string
      username:
        type: string
      bio:
        type: string
      image:
        type: string
  UpdateUserRequest:
    type: object
    properties:
      user:
        $ref: "#/definitions/UpdateUser"
    required:
      - user
...
  GenericErrorModel:
    type: object
    properties:
      errors:
        type: object
        properties:
          body:
            type: array
            items:
              type: string
        required:
          - body
    required:
      - errors